# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API Server for University Student Admission System",
        "title": "DBMS Project API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/program/requirements": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Checks if the authenticated student has fulfilled all required profile fields for a specific program.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Requirements"
                ],
                "summary": "Get program requirements status",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Program ID",
                        "name": "program_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Requirements status response",
                        "schema": {
                            "$ref": "#/definitions/internal_handler_rest.ProgramRequirementsResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid program_id",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "405": {
                        "description": "Method not allowed",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Failed to fetch requirements",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "internal_handler_rest.ProgramRequirementsResponse": {
            "type": "object",
            "properties": {
                "is_ready_to_apply": {
                    "type": "boolean"
                },
                "missing_fields": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "program_id": {
                    "type": "integer"
                },
                "required_fields": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/internal_handler_rest.RequiredFieldStatus"
                    }
                }
            }
        },
...
        "internal_handler_rest.RequiredFieldStatus": {
            "type": "object",
            "properties": {
                "field_name": {
                    "type": "string"
                },
                "is_provided": {
                    "type": "boolean"
                },
                "value": {
                    "description": "Null if not filled yet",
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer \" followed by your JWT token",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}